# IGNITE-13883 .NET: Simplify and optimize TryReadSystemType