--- a/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Binary/BinarySystemTypeReadBenchmark.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Binary/BinarySystemTypeReadBenchmark.cs
@@ -15,6 +15,7 @@
  * limitations under the License.
  */
 
+// ReSharper disable RedundantCast
 namespace Apache.Ignite.BenchmarkDotNet.Binary
 {
     using System;