--- a/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Binary/BinarySystemTypeReadBenchmark.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Binary/BinarySystemTypeReadBenchmark.cs
@@ -27,6 +27,7 @@ namespace Apache.Ignite.BenchmarkDotNet.Binary
     /// <summary>
     /// System type reading benchmark. Checks <see cref="BinarySystemHandlers.TryReadSystemType{T}"/> performance.
     /// </summary>
+    [MemoryDiagnoser]
     public class BinarySystemTypeReadBenchmark
     {
         /** */