--- a/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Binary/BinarySystemTypeReadBenchmark.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Binary/BinarySystemTypeReadBenchmark.cs
@@ -26,6 +26,10 @@ namespace Apache.Ignite.BenchmarkDotNet.Binary
 
     /// <summary>
     /// System type reading benchmark. Checks <see cref="BinarySystemHandlers.TryReadSystemType{T}"/> performance.
+    /// <para />
+    /// | Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
+    /// |------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
+    /// |   Read | 22.10 us | 0.107 us | 0.100 us | 1.5869 |     - |     - |   9.79 KB |
     /// </summary>
     [MemoryDiagnoser]
     public class BinarySystemTypeReadBenchmark