--- a/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Binary/BinarySystemTypeReadBenchmark.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Binary/BinarySystemTypeReadBenchmark.cs
@@ -19,10 +19,12 @@
 namespace Apache.Ignite.BenchmarkDotNet.Binary
 {
     using System;
+    using System.IO;
     using Apache.Ignite.Core.Binary;
     using Apache.Ignite.Core.Impl.Binary;
     using Apache.Ignite.Core.Impl.Memory;
     using global::BenchmarkDotNet.Attributes;
+    using BinaryReader = Apache.Ignite.Core.Impl.Binary.BinaryReader;
 
     /// <summary>
     /// System type reading benchmark. Checks <see cref="BinarySystemHandlers.TryReadSystemType{T}"/> performance.
@@ -41,13 +43,13 @@ namespace Apache.Ignite.BenchmarkDotNet.Binary
         private static readonly Guid Guid = Guid.NewGuid();
 
         /** */
-        private readonly Marshaller _marsh = new Marshaller(new BinaryConfiguration {ForceTimestamp = true});
+        private static readonly Marshaller Marsh = new Marshaller(new BinaryConfiguration {ForceTimestamp = true});
 
         /** */
-        private readonly PlatformMemoryManager _memMgr = new PlatformMemoryManager(1024);
+        private static readonly PlatformMemoryManager MemMgr = new PlatformMemoryManager(1024);
 
         /** */
-        private IPlatformMemory _mem;
+        private BinaryReader _reader;
 
         /// <summary>
         /// Sets up the benchmark.
@@ -55,9 +57,9 @@ namespace Apache.Ignite.BenchmarkDotNet.Binary
         [GlobalSetup]
         public void Setup()
         {
-            _mem = _memMgr.Allocate();
-            var stream = _mem.GetStream();
-            var writer = _marsh.StartMarshal(stream);
+            var mem = MemMgr.Allocate();
+            var stream = mem.GetStream();
+            var writer = Marsh.StartMarshal(stream);
 
             writer.Write(true);
             writer.Write('i');
@@ -84,37 +86,38 @@ namespace Apache.Ignite.BenchmarkDotNet.Binary
             writer.Write(new Guid?[] {Guid});
 
             stream.SynchronizeOutput();
+
+            _reader = Marsh.StartUnmarshal(stream);
         }
 
         [Benchmark]
         public void Read()
         {
-            var stream = _mem.GetStream();
-            var reader = _marsh.StartUnmarshal(stream);
+            _reader.Stream.Seek(0, SeekOrigin.Begin);
 
-            reader.ReadObject<bool>();
-            reader.ReadObject<byte>();
-            reader.ReadObject<char>();
-            reader.ReadObject<short>();
-            reader.ReadObject<int>();
-            reader.ReadObject<long>();
-            reader.ReadObject<float>();
-            reader.ReadObject<double>();
-            reader.ReadObject<decimal>();
-            reader.ReadObject<DateTime>();
-            reader.ReadObject<Guid>();
+            _reader.ReadObject<bool>();
+            _reader.ReadObject<byte>();
+            _reader.ReadObject<char>();
+            _reader.ReadObject<short>();
+            _reader.ReadObject<int>();
+            _reader.ReadObject<long>();
+            _reader.ReadObject<float>();
+            _reader.ReadObject<double>();
+            _reader.ReadObject<decimal>();
+            _reader.ReadObject<DateTime>();
+            _reader.ReadObject<Guid>();
 
-            reader.ReadObject<bool[]>();
-            reader.ReadObject<char[]>();
-            reader.ReadObject<byte[]>();
-            reader.ReadObject<short[]>();
-            reader.ReadObject<int[]>();
-            reader.ReadObject<long[]>();
-            reader.ReadObject<float[]>();
-            reader.ReadObject<double[]>();
-            reader.ReadObject<decimal?[]>();
-            reader.ReadObject<DateTime?[]>();
-            reader.ReadObject<Guid?[]>();
+            _reader.ReadObject<bool[]>();
+            _reader.ReadObject<char[]>();
+            _reader.ReadObject<byte[]>();
+            _reader.ReadObject<short[]>();
+            _reader.ReadObject<int[]>();
+            _reader.ReadObject<long[]>();
+            _reader.ReadObject<float[]>();
+            _reader.ReadObject<double[]>();
+            _reader.ReadObject<decimal?[]>();
+            _reader.ReadObject<DateTime?[]>();
+            _reader.ReadObject<Guid?[]>();
         }
     }
 }