--- a/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Program.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Program.cs
@@ -17,7 +17,7 @@
 
 namespace Apache.Ignite.BenchmarkDotNet
 {
-    using Apache.Ignite.BenchmarkDotNet.ThinClient;
+    using Apache.Ignite.BenchmarkDotNet.Binary;
     using global::BenchmarkDotNet.Running;
 
     /// <summary>
@@ -30,7 +30,7 @@ namespace Apache.Ignite.BenchmarkDotNet
         /// </summary>
         public static void Main()
         {
-            BenchmarkRunner.Run<ThinClientServicesBenchmark>();
+            BenchmarkRunner.Run<BinarySystemTypeReadBenchmark>();
         }
     }
 }