--- a/modules/platforms/dotnet/Apache.Ignite.Core.Tests/Services/Model.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Core.Tests/Services/Model.cs
@@ -94,7 +94,7 @@ namespace org.apache.ignite.platform.model
     public class Account
     {
         public String Id { get; set; }
-        
+
         public int Amount { get; set; }
 
         protected bool Equals(Account other)
@@ -120,6 +120,7 @@ namespace org.apache.ignite.platform.model
     /// <summary>
     /// A enum is a clone of Java class User with the same namespace.
     /// </summary>
+    // ReSharper disable once InconsistentNaming
     public enum ACL
     {
         Allow, Deny
@@ -139,9 +140,9 @@ namespace org.apache.ignite.platform.model
     public class User
     {
         public int Id { get; set; }
-        
+
         public ACL Acl { get; set; }
-        
+
         public Role Role { get; set; }
-    }    
+    }
 }